+++ src/BenjolinOsc.cpp
@@ -6,6 +6,7 @@
 #include "BenjolinModule.hpp"
 
 #define LIGHT_ON 1.f
+#define LIGHT_DIM 0.2f
 #define LIGHT_OFF 0.02f
 
 struct BenjolinOsc : BenjolinModule {
@@ -44,6 +45,7 @@ struct BenjolinOsc : BenjolinModule {
   enum LightId {
     CHAOS_LIGHT,
     DOUBLE_LIGHT,
+    ENUMS(RUNGLER_LIGHT,8),
     LIGHTS_LEN
   };
 
@@ -198,6 +200,16 @@ struct BenjolinOsc : BenjolinModule {
       else
         outA = osc*5.f;
     }
+
+    lights[RUNGLER_LIGHT+0].setBrightnessSmooth(asr&1   ? LIGHT_DIM : LIGHT_OFF, args.sampleTime);
+    lights[RUNGLER_LIGHT+1].setBrightnessSmooth(asr&2   ? (dacMode?LIGHT_DIM:LIGHT_ON) : 0.f, args.sampleTime);
+    lights[RUNGLER_LIGHT+2].setBrightnessSmooth(asr&4   ? LIGHT_DIM : 0.f, args.sampleTime);
+    lights[RUNGLER_LIGHT+3].setBrightnessSmooth(asr&8   ? (dacMode?LIGHT_DIM:LIGHT_ON) : 0.f, args.sampleTime);
+    lights[RUNGLER_LIGHT+4].setBrightnessSmooth(asr&16  ? LIGHT_DIM : 0.f, args.sampleTime);
+    lights[RUNGLER_LIGHT+5].setBrightnessSmooth(asr&32  ? (dacMode?LIGHT_ON:LIGHT_DIM) : 0.f, args.sampleTime);
+    lights[RUNGLER_LIGHT+6].setBrightnessSmooth(asr&64  ? LIGHT_ON : 0.f, args.sampleTime);
+    lights[RUNGLER_LIGHT+7].setBrightnessSmooth(asr&128 ? (dacMode?LIGHT_ON:LIGHT_DIM) : 0.f, args.sampleTime);
+
     trig = 0;
     if (clockTrig.isHigh() != oldTrig){
       trig = clockTrig.isHigh() ? 1 : -1;
@@ -342,7 +354,7 @@ struct BenjolinOscWidget : VenomWidget {
     setVenomPanel("BenjolinOsc");
     addParam(createLockableParamCentered<RoundBlackKnobLockable>(Vec(24.865,96.097), module, BenjolinOsc::FREQ1_PARAM));
     addParam(createLockableParamCentered<RoundBlackKnobLockable>(Vec(63.288,96.097), module, BenjolinOsc::FREQ2_PARAM));
-    addParam(createLockableParamCentered<OverSwitch>(Vec(120.923,76.204), module, BenjolinOsc::OVER_PARAM));
+    addParam(createLockableParamCentered<OverSwitch>(Vec(120.923,68.204), module, BenjolinOsc::OVER_PARAM));
     addParam(createLockableParamCentered<RoundSmallBlackKnobLockable>(Vec(24.865,148.33), module, BenjolinOsc::RUNG1_PARAM));
     addParam(createLockableParamCentered<RoundSmallBlackKnobLockable>(Vec(63.288,148.33), module, BenjolinOsc::RUNG2_PARAM));
     addParam(createLockableParamCentered<RoundBlackKnobLockable>(Vec(120.923,148.33), module, BenjolinOsc::PATTERN_PARAM));
@@ -362,6 +374,14 @@ struct BenjolinOscWidget : VenomWidget {
     addOutput(createOutputCentered<MonoPort>(Vec(63.288,330.368), module, BenjolinOsc::PULSE2_OUTPUT));
     addOutput(createOutputCentered<MonoPort>(Vec(101.712,330.368), module, BenjolinOsc::PWM_OUTPUT));
     addOutput(createOutputCentered<MonoPort>(Vec(140.135,330.368), module, BenjolinOsc::RUNG_OUTPUT));
+
+    {
+      int i;
+      float x;
+      for (i=0, x=93.066f; i<8; i++, x+=7.959f){
+        addChild(createLightCentered<SmallSimpleLight<YellowLight>>(Vec(x,90.137f), module, BenjolinOsc::RUNGLER_LIGHT+i));
+      }
+    }
   }
 
   void appendContextMenu(Menu* menu) override {